third_party/blink/web_tests/fast/dom/SelectorAPI/caseID-almost-strict.html

<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<title>fast/dom/SelectorAPI/caseID-almost-strict.html</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
    <div style="display: none">
    <span id="lower1">lower 1</span><br>
    <span id="lower2">lower 2</span><br>
    <span id="UPPER1">UPPER 1</span><br>
    <span id="UPPER2">UPPER 2</span><br>
    </div>
    <script>
      setup({ single_test: true });
      assert_equals(document.querySelector('#lower1').textContent, 'lower 1');
      assert_equals(document.querySelector('#LOWER2'), null);
      assert_equals(document.querySelector('#UPPER1').textContent, 'UPPER 1');
      assert_equals(document.querySelector('#upper2'), null);

      assert_true(document.getElementById('lower1').webkitMatchesSelector('#lower1'));
      assert_false(document.getElementById('lower2').webkitMatchesSelector('#LOWER2'));
      assert_true(document.getElementById('UPPER1').webkitMatchesSelector('#UPPER1'));
      assert_false(document.getElementById('UPPER2').webkitMatchesSelector('#upper2'));
      done();
    </script>
</body>
</html>
